requirements.txt
numpy>=1.24
scipy
pydub
pandas>=2.0
scikit-learn>=1.3
librosa>=0.10.1
soundfile>=0.12.1
tqdm>=4.66
PyYAML>=6.0
matplotlib>=3.8